Removed length assertion
Was mainly there for testing how often precision fails on me, and by how much. It does throw an exception sometimes, but the deviation is not enough to show-up in `std::to_string( Result.Length())`, and I'm too lazy to get into how it prints doubles and if there is a deviation 15 points after the dot that gets rounded.